Split out components, remove Feeds
changed target to pusher platform

Clean up the scripts, remove feeds.ts
