+++ package.json
@@ -13,15 +13,15 @@
     "url": "git+https://github.com/pusher/pusher-platform-js.git"
   },
   "devDependencies": {
-    "ts-loader": "^1.3.3",
-    "typescript": "^2.0.7",
-    "webpack": "^1.9.5"
+    "ts-loader": "^2.1.0",
+    "typescript": "^2.3.2",
+    "webpack": "^2.5.1"
   },
   "scripts": {
-    "build": "rm -rf lib target && node_modules/webpack/bin/webpack.js --config=webpack.js",
-    "test": "node_modules/webpack/bin/webpack.js --config=webpack.js && node test/test.js"
+    "build": "rm -rf lib target && node_modules/webpack/bin/webpack.js",
+    "test": "node_modules/webpack/bin/webpack.js && node test/test.js "
   },
-  "version": "0.0.5",
+  "version": "0.1.0",
   "dependencies": {
     "jwt-simple": "^0.5.1"
   }
