+++ tsconfig.json
@@ -1,11 +1,8 @@
 {
   "compilerOptions": {
+    "target": "ES5",
+    "sourceMap": true,
     "module": "commonjs",
-    "outDir": "lib/",
-    "target": "es5",
-    "lib": ["es6", "dom"]
-  },
-  "files": [
-    "src/pusher-platform.ts"
-  ]
+    "lib": ["ES6", "dom"]
+  }
 }
